+++ map.h
@@ -9,10 +9,10 @@ typedef void cleanupfunction(breakpoint*);
 typedef struct UnorderedMap map;
 
 map* map_init(uint32_t size, cleanupfunction* cf);
-bool map_insert(map* ht, long key, breakpoint* obj);
-void* map_lookup(map* ht, long key);
-bool map_it_exsists(map* ht, long key);
-void* map_delete(map* ht, long key);
+bool map_insert(map* ht, uintptr_t key, breakpoint* obj);
+void* map_lookup(map* ht, uintptr_t key);
+bool map_it_exsists(map* ht, uintptr_t key);
+void* map_delete(map* ht, uintptr_t key);
 void map_free(map* ht);
 
 #endif
